Moved WizardPanel to wizard.py.
