+++ phy/cluster/manual/session.py
@@ -34,7 +34,7 @@ from .view_model import (WaveformViewModel,
                          CorrelogramViewModel,
                          TraceViewModel,
                          )
-from .wizard import Wizard, _best_clusters
+from .wizard import Wizard, WizardPanel, _best_clusters
 
 
 #------------------------------------------------------------------------------
@@ -384,161 +384,6 @@ class FeatureMasks(StoreItem):
             self._assign(up)
         # Compute the extra fields for the new clusters.
         self._store_extra_fields(up.added)
-
-
-#------------------------------------------------------------------------------
-# Wizard panel
-#------------------------------------------------------------------------------
-
-_PANEL_HTML = """
-<div class="control-panel">
-<div class="best">
-    <div class="id">{best}</div>
-    <div class="progress">
-        <progress value="{best_progress:d}" max="100"></progress>
-    </div>
-</div>
-<div class="match">
-    <div class="id">{match}</div>
-    <div class="progress">
-        <progress value="{match_progress:d}" max="100"></progress>
-    </div>
-</div>
-</div>"""
-
-
-_PANEL_CSS = """
-html, body, div {
-    background-color: black;
-}
-
-.control-panel {
-    background-color: black;
-    color: white;
-    font-weight: bold;
-    font-size: 24pt;
-    padding: 10px;
-    text-align: center
-}
-
-.control-panel > div {
-    display: inline-block;
-    margin: 0 auto;
-}
-
-.control-panel .best {
-    margin-right: 20px;
-    color: rgb(102, 194, 165);
-}
-
-.control-panel .match {
-    color: rgb(252, 141, 98);
-}
-
-.control-panel > div .id {
-    margin: 10px 0 20px 0;
-    height: 40px;
-    text-align: center;
-    vertical-align: middle;
-}
-
-.control-panel progress[value] {
-    width: 200px;
-}
-
-"""
-
-
-def _wizard_panel_html(best=None,
-                       best_progress=None,
-                       match=None,
-                       match_progress=None,
-                       ):
-    out = '<style>' + _PANEL_CSS + '</style>\n'
-    out += _PANEL_HTML.format(best=best,
-                              best_progress=best_progress,
-                              match=match,
-                              match_progress=match_progress,
-                              )
-    return out
-
-
-class WizardPanel(object):
-    def __init__(self):
-        self._best = None
-        self._match = None
-        self._best_index = 0
-        self._match_index = 0
-        self._best_count = 0
-        self._match_count = 0
-
-    @property
-    def best(self):
-        return self._best
-
-    @best.setter
-    def best(self, value):
-        self._best = value
-
-    @property
-    def match(self):
-        return self._match
-
-    @match.setter
-    def match(self, value):
-        self._match = value
-
-    @property
-    def best_index(self):
-        return self._best_index
-
-    @best_index.setter
-    def best_index(self, value):
-        self._best_index = value
-
-    @property
-    def best_count(self):
-        return self._best_count
-
-    @best_count.setter
-    def best_count(self, value):
-        self._best_count = value
-
-    @property
-    def match_index(self):
-        return self._match_index
-
-    @match_index.setter
-    def match_index(self, value):
-        self._match_index = value
-
-    @property
-    def match_count(self):
-        return self._match_count
-
-    @match_count.setter
-    def match_count(self, value):
-        self._match_count = value
-
-    def _progress(self, value, maximum):
-        if maximum <= 1:
-            return 1
-        return int(100 * value / float(maximum - 1))
-
-    @property
-    def html(self):
-        bp = self._progress(self.best_index, self.best_count)
-        mp = self._progress(self.match_index, self.match_count)
-        return _wizard_panel_html(best=self.best
-                                  if self.best is not None else '',
-                                  match=self.match
-                                  if self.match is not None else '',
-                                  best_progress=bp,
-                                  match_progress=mp,
-                                  )
-
-    def _repr_html_(self):
-        return self.html
 
 
 #------------------------------------------------------------------------------
